+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support to copy CAPI cluster's certificates
+
 ## [1.24.6] - 2021-03-02
 
 ### Fixed